+++ backend/registry/src/lib.rs
@@ -3,6 +3,9 @@
 //! This data is mostly accessed during problem generation, and since data about problems
 //! won't change during runtime (in production), this is done to minimize DB hits (100 problems
 //! would cause AT LEAST 100 DB hits).
+mod split_strings;
+use split_strings::*;
+
 use anyhow::{Context, Result};
 use db::{self, PrefixEntry, ProblemEntry};
 use std::collections::HashMap;
@@ -18,11 +21,11 @@ pub enum RegistryError {
     RegistryMutexIsPoisoned { registry: String },
 }
 
-/// A map between a `Problem's` id in the DB and a [`ProblemEntry`] struct.
+/// A map between a `Problem's` id in the DB and a [`SplitProblemEntry`] struct.
 ///
 /// The `ProblemEntry` contains all the information about the problem, like questions for each
 /// language, its difficulty, etc.
-pub static PROBLEM_DATA: LazyLock<RwLock<HashMap<i32, ProblemEntry>>> =
+pub static PROBLEM_DATA: LazyLock<RwLock<HashMap<i32, SplitProblemEntry>>> =
     LazyLock::new(|| RwLock::new(HashMap::new()));
 
 /// A map between a `Prefix's` id in the DB and a [`PrefixEntry`] struct.
@@ -39,7 +42,7 @@ pub async fn load_problem_data() -> Result<()> {
             .write()
             .ok()
             .context("Failed to write to PROBLEM_DATA")?
-            .insert(problem.id, problem);
+            .insert(problem.id, problem.try_into()?);
     }
     Ok(())
 }
@@ -74,13 +77,24 @@ pub fn get_prefix_data(id: i32) -> Result<PrefixEntry> {
 
 /// Used in problems to retrieve questions, answers, etc.
 pub fn get_problem_data(id: i32) -> Result<ProblemEntry> {
-    let problem = PROBLEM_DATA
-        .read()
+    match PROBLEM_DATA
+        .write()
         .map_err(|_| RegistryError::RegistryMutexIsPoisoned {
             registry: "PROBLEM_DATA".to_string(),
         })?
-        .get(&id)
-        .cloned()
-        .ok_or(RegistryError::ProblemNotFound { id })?;
-    Ok(problem)
+        .get_mut(&id)
+        .ok_or(RegistryError::ProblemNotFound { id })?
+    {
+        SplitProblemEntry::Single(entry) => Ok(entry.clone()),
+        SplitProblemEntry::Multiple {
+            entry,
+            latest_index,
+            split_texts,
+        } => {
+            *latest_index = (*latest_index + 1) % split_texts.iter().len();
+            let mut new_entry = entry.clone();
+            new_entry.translations = split_texts[*latest_index].clone();
+            Ok(new_entry)
+        }
+    }
 }